Add the Withdraw instruction (#5)
Adds the Withdraw instruction, which allows users to withdraw from a market that is invalid and finalized. If the market can be auto-finalized (i.e. `market.finalize(ts)` would return true), this instruction can also finalize the outcome, thus removing the need to an extra call to the UpdateStatus instruction.

Also removes the finalized check from the InitializeUserPosition instruction, which is unnecessary and helps simplify testing.